chore: refactor type
